src/test/java/com/swirlds/regression/GitInfoTest.java
/*
 * (c) 2016-2019 Swirlds, Inc.
 *
 * This software is the confidential and proprietary information of
 * Swirlds, Inc. ("Confidential Information"). You shall not
 * disclose such Confidential Information and shall use it only in
 * accordance with the terms of the license agreement you entered into
 * with Swirlds.
 *
 * SWIRLDS MAKES NO REPRESENTATIONS OR WARRANTIES ABOUT THE SUITABILITY OF
 * THE SOFTWARE, EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED
 * TO THE IMPLIED WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
 * PARTICULAR PURPOSE, OR NON-INFRINGEMENT. SWIRLDS SHALL NOT BE LIABLE FOR
 * ANY DAMAGES SUFFERED BY LICENSEE AS A RESULT OF USING, MODIFYING OR
 * DISTRIBUTING THIS SOFTWARE OR ITS DERIVATIVES.
 */

package com.swirlds.regression;

import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import java.io.File;
import java.io.IOException;
import java.net.URI;
import java.net.URISyntaxException;
import java.nio.file.Files;
import java.nio.file.Path;
import java.nio.file.Paths;
import java.util.ArrayList;
import java.util.List;
import java.util.stream.Collectors;
import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.junit.jupiter.api.Assertions.assertTrue;

public class GitInfoTest {

	@Test
	@DisplayName("Test Git Version To File")
	public void testGitVersionToFile() {
		GitInfo gi = new GitInfo();
		gi.gitVersionToFile();
		File procLog = new File("gitLog.log");
		assertNotNull(procLog);
		assertTrue(procLog.exists());
		assertTrue(procLog.isFile());
	}

	@Test
	@DisplayName("Test Git Info")
	public void testGitInfo() {
		GitInfo gi = new GitInfo();
		gi.gitVersionInfo();
		String info = gi.getGitInfo(true);
		//System.out.println("*" + info + "*");
		String date = gi.getLastCommitDate();
		//System.out.println(date);
	}

}
